core/requirements.txt
python-engineio==3.14.2
python-socketio==4.6.1
eventlet
pynput
requests
